requirements.txt

requests==2.19.1

# database
psycopg2==2.7.5

Django==1.11.18

# django rest framework
djangorestframework==3.8.2

dj-database-url==0.3.0
python-decouple==3
whitenoise